Fix unnecessary shaded API, shrinks jar from 6 to 2.6 MB. Closes GH-28
The Bukkit API was specified with compile scope, causing it to be shaded
into the plugin jar by default. Requested change to Bukkit/SamplePlugin
here: https://github.com/Bukkit/SamplePlugin/pull/16 - to scope
provided, this is intended for dependencies provided at runtime (i.e.,
Bukkit) so removes all the unnecessary deps in the jar, ebean etc.

Remainder of this commit is coping with this fix: cleaning up the
unneeded relocations and exclusions, and breaking the unintentional
dependency on Google Commons where Charset was used although it is not
needed and available in java.nio. Netty is still shaded for server
software version independence, but even with it included and the
NetCraft binary, end result is plugin shrinks from 6.0 to 2.6 MB, well
under the Spigot Resources limit of 4 MB.