Remove blocks_to_web in favor of blocks_to_web_override. Closes GH-62
The blocks_to_web configuration option used to contain a map from Bukkit
material names to web client IDs, set to a default. But the problem with
this is that when I added or changed blocks, since the default was
written out, the old default would remain when server administrators
upgrade from older versions of this plugin.

To solve this, remove blocks_to_web and instead add
blocks_to_web_override, defaulting to empty. The translations I update
are instead stored internally in the plugin, not written out, so they
can be changed freely, and blocks_to_override left empty unless the
server administrator specifically wants to change something.